refactor(server): import groqCleanupProviderOptions from its canonical module
Drop the backward-compat re-export from post-process.ts and point the test
and benchmark scripts at ./llm/registry.js, where the Groq provider
descriptor now owns groqCleanupProviderOptions.
